Lorentz.ADT

Documentation

type HasField dt fname = (InstrGetFieldC dt fname, InstrSetFieldC dt fname)

Allows field access and modification.

type HasFieldOfType dt fname fieldTy = (HasField dt fname, GetFieldType dt fname ~ fieldTy)

Like HasField, but allows constrainting field type.

type family HasFieldsOfType (dt :: Type) (fs :: [NamedField]) :: Constraint where ...

Shortcut for multiple HasFieldOfType constraints.

Equations

HasFieldsOfType _ '[] = ()

HasFieldsOfType dt ((n := ty) ': fs) = (HasFieldOfType dt n ty, HasFieldsOfType dt fs)

data NamedField

A pair of field name and type.

Constructors

NamedField Symbol Type

type (:=) n ty = NamedField n ty infixr 0

toField :: forall dt name st. InstrGetFieldC dt name => Label name -> (dt & st) :-> (GetFieldType dt name & st)

Extract a field of a datatype replacing the value of this datatype with the extracted field.

For this and the following functions you have to specify field name which is either record name or name attached with (:!) operator.

toFieldNamed :: forall dt name st. InstrGetFieldC dt name => Label name -> (dt & st) :-> ((name :! GetFieldType dt name) & st)

Like toField, but leaves field named.

getField :: forall dt name st. InstrGetFieldC dt name => Label name -> (dt & st) :-> (GetFieldType dt name & (dt ': st))

Extract a field of a datatype, leaving the original datatype on stack.

getFieldNamed :: forall dt name st. InstrGetFieldC dt name => Label name -> (dt & st) :-> ((name :! GetFieldType dt name) & (dt ': st))

Like getField, but leaves field named.

setField :: forall dt name st. InstrSetFieldC dt name => Label name -> (GetFieldType dt name ': (dt ': st)) :-> (dt ': st)

Set a field of a datatype.

modifyField :: forall dt name st. (InstrGetFieldC dt name, InstrSetFieldC dt name) => Label name -> (forall st0. (GetFieldType dt name ': st0) :-> (GetFieldType dt name ': st0)) -> (dt & st) :-> (dt & st)

Apply given modifier to a datatype field.

construct :: forall dt st. (InstrConstructC dt, RMap (ConstructorFieldTypes dt)) => Rec (FieldConstructor st) (ConstructorFieldTypes dt) -> st :-> (dt & st)

Make up a datatype. You provide a pack of individual fields constructors.

Each element of the accepted record should be an instruction wrapped with fieldCtor function. This instruction will have access to the stack at the moment of calling construct. Instructions have to output fields of the built datatype, one per instruction; instructions order is expected to correspond to the order of fields in the datatype.

constructT :: forall dt fctors st. (InstrConstructC dt, RMap (ConstructorFieldTypes dt), fctors ~ Rec (FieldConstructor st) (ConstructorFieldTypes dt), RecFromTuple fctors) => IsoRecTuple fctors -> st :-> (dt & st)

Version of construct which accepts tuple of field constructors.

fieldCtor :: HasCallStack => (st :-> (f & st)) -> FieldConstructor st f

Lift an instruction to field constructor.

wrap_ :: forall dt name st. InstrWrapC dt name => Label name -> AppendCtorField (GetCtorField dt name) st :-> (dt & st)

Wrap entry in constructor. Useful for sum types.

case_ :: forall dt out inp. (InstrCaseC dt, RMap (CaseClauses dt)) => Rec (CaseClauseL inp out) (CaseClauses dt) -> (dt & inp) :-> out

Pattern match on the given sum type.

You have to provide a Rec containing case branches. To construct a case branch use /-> operator.

caseT :: forall dt out inp clauses. CaseTC dt out inp clauses => IsoRecTuple clauses -> (dt & inp) :-> out

Like case_, accepts a tuple of clauses, which may be more convenient.

If user is experiencing problems with wierd errors about tuples while using this function, he should take look at Instances and ensure that his tuple isn't bigger than generated instances, if so, he should probably extend number of generated instances.

unwrapUnsafe_ :: forall dt name st. InstrUnwrapC dt name => Label name -> (dt & st) :-> (CtorOnlyField name dt ': st)

Wrap entry in constructor. Useful for sum types.

type CaseTC dt out inp clauses = (InstrCaseC dt, RMap (CaseClauses dt), RecFromTuple clauses, clauses ~ Rec (CaseClauseL inp out) (CaseClauses dt))

class CaseArrow name body clause | clause -> name, clause -> body where

Provides "case" arrow which works on different wrappers for clauses.

Methods

(/->) :: Label name -> body -> clause infixr 0

Lift an instruction to case clause.

You should write out constructor name corresponding to the clause explicitly. Prefix constructor name with "c" letter, otherwise your label will not be recognized by Haskell parser. Passing constructor name can be circumvented but doing so is not recomended as mentioning contructor name improves readability and allows avoiding some mistakes.

data CaseClauseL (inp :: [Type]) (out :: [Type]) (param :: CaseClauseParam) where

Lorentz analogy of CaseClause, it works on plain Type types.

Constructors

CaseClauseL :: (AppendCtorField x inp :-> out) -> CaseClauseL inp out (CaseClauseParam ctor x)

type InstrConstructC dt = (GenericIsoValue dt, GInstrConstruct (Rep dt))

Constraint for instrConstruct.

type ConstructorFieldTypes dt = GFieldTypes (Rep dt)

Types of all fields in a datatype.

Useful re-exports

data Rec (a :: u -> Type) (b :: [u]) :: forall u. (u -> Type) -> [u] -> Type where

A record is parameterized by a universe u, an interpretation f and a list of rows rs. The labels or indices of the record are given by inhabitants of the kind u; the type of values at any label r :: u is given by its interpretation f r :: *.

Constructors

RNil :: forall u (a :: u -> Type) (b :: [u]). Rec a ([] :: [u])

(:&) :: forall u (a :: u -> Type) (b :: [u]) (r :: u) (rs :: [u]). !(a r) -> !(Rec a rs) -> Rec a (r ': rs) infixr 7

type (:!) (name :: Symbol) a = NamedF Identity a name

Infix notation for the type of a named parameter.

type (:?) (name :: Symbol) a = NamedF Maybe a name

Infix notation for the type of an optional named parameter.

arg :: Name name -> (name :! a) -> a

arg unwraps a named parameter with the specified name. One way to use it is to match on arguments with -XViewPatterns:

fn (arg #t -> t) (arg #f -> f) = ...

This way, the names of parameters can be inferred from the patterns: no type signature for fn is required. In case a type signature for fn is provided, the parameters must come in the same order:

fn :: "t" :! Integer -> "f" :! Integer -> ...
fn (arg #t -> t) (arg #f -> f) = ... -- ok
fn (arg #f -> f) (arg #t -> t) = ... -- does not typecheck

argDef :: Name name -> a -> (name :? a) -> a

A variation of arg for optional arguments. Requires a default value to handle the case when the optional argument was omitted:

fn (argDef #answer 42 -> ans) = ...

In case you want to get a value wrapped in Maybe instead, use argF or ArgF.

argF :: Name name -> NamedF f a name -> f a

argF is similar to arg: it unwraps a named parameter with the specified name. The difference is that the result of argF is inside an arity wrapper, which is Identity for normal parameters and Maybe for optional parameters.