Lorentz.Address

Description

This module introduces several types for safe work with address and contract types. All available types for that are represented in the following table:

Type 	 Type safe? 	 What it refers to 	 Michelson reflection

Address 	 No 	 Whole contract 	 address

EpAddress 	 No 	 Entrypoint 	 address

TAddress 	 Yes 	 Whole contract 	 address

FutureContract 	 Yes 	 Entrypoint 	 address

ContractRef 	 Yes 	 Entrypoint 	 contract

This module also provides functions for converting between these types in Haskell and Michelson worlds. In the latter you can additionally use coercions and dedicated instructions from Lorentz.Instr.

Documentation

newtype TAddress p

Address which remembers the parameter type of the contract it refers to.

It differs from Michelson's contract type because it cannot contain entrypoint, and it always refers to entire contract parameter even if this contract has explicit default entrypoint.

Constructors

TAddress

Fields

unTAddress :: Address

newtype FutureContract arg

Address associated with value of contract arg type.

Places where ContractRef can appear are now severely limited, this type gives you type-safety of ContractRef but still can be used everywhere. This type is not a full-featured one rather a helper; in particular, once pushing it on stack, you cannot return it back to Haskell world.

Note that it refers to an entrypoint of the contract, not just the contract as a whole. In this sense this type differs from TAddress.

Unlike with ContractRef, having this type you still cannot be sure that the referred contract exists and need to perform a lookup before calling it.

Constructors

FutureContract

Fields

unFutureContract :: ContractRef arg

Conversions

callingTAddress :: forall cp mname. NiceParameterFull cp => TAddress cp -> EntryPointRef mname -> ContractRef (GetEntryPointArgCustom cp mname)

Turn TAddress to ContractRef in Haskell world.

This is an analogy of address to contract convertion in Michelson world, thus you have to supply an entrypoint (or call the default one explicitly).

callingDefTAddress :: forall cp. NiceParameterFull cp => TAddress cp -> ContractRef (GetDefaultEntryPointArg cp)

Specification of callTAddress to call the default entrypoint.

class ToAddress a where

Convert something to Address in Haskell world.

Use this when you want to access state of the contract and are not interested in calling it.

Methods

toAddress :: a -> Address

class ToTAddress (cp :: Type) (a :: Type) where

Convert something referring to a contract (not specific entrypoint) to TAddress in Haskell world.

Methods

toTAddress :: a -> TAddress cp

type ToTAddress_ cp addr = (ToTAddress cp addr, ToT addr ~ ToT Address)

Something coercible to 'TAddress cp'.

toTAddress_ :: forall cp addr s. ToTAddress_ cp addr => (addr ': s) :-> (TAddress cp ': s)

Cast something appropriate to TAddress.

class ToContractRef (cp :: Type) (contract :: Type) where

Convert something to ContractRef in Haskell world.

Methods

toContractRef :: HasCallStack => contract -> ContractRef cp

class FromContractRef (cp :: Type) (contract :: Type) where

Convert something from ContractAddr in Haskell world.

Methods

fromContractRef :: ContractRef cp -> contract

convertContractRef :: forall cp contract2 contract1. (ToContractRef cp contract1, FromContractRef cp contract2) => contract1 -> contract2

Re-exports

data Address

Data type corresponding to address structure in Tezos.

data EpAddress

Address with optional entrypoint name attached to it. TODO: come up with better name?

Constructors

EpAddress

Fields

eaAddress :: Address

Address itself

eaEntryPoint :: EpName

Entrypoint name (might be empty)

data ContractRef arg

Since Contract name is used to designate contract code, lets call analogy of TContract type as follows.

Note that type argument always designates an argument of entrypoint. If a contract has explicit default entrypoint (and no root entrypoint), ContractRef referring to it can never have the entire parameter as its type argument.

Constructors

ContractRef

Fields

crAddress :: Address

crEntryPoint :: SomeEntryPointCall arg

coerceContractRef :: ToT a ~ ToT b => ContractRef a -> ContractRef b

Replace type argument of ContractAddr with isomorphic one.