Lorentz.Base

Description

Foundation of Lorentz development.

Documentation

newtype (inp :: [Type]) :-> (out :: [Type]) infixr 1

Alias for instruction which hides inner types representation via T.

Constructors

LorentzInstr

Fields

unLorentzInstr :: RemFail Instr (ToTs inp) (ToTs out)

type (%>) = (:->) infixr 1

Alias for :->, seems to make signatures more readable sometimes.

Let's someday decide which one of these two should remain.

type (&) (a :: Type) (b :: [Type]) = a ': b infixr 2

(#) :: (a :-> b) -> (b :-> c) -> a :-> c infixl 8

(##) :: (a :-> b) -> (b :-> c) -> a :-> c

Version of # which performs some optimizations immediately.

pattern I :: Instr (ToTs inp) (ToTs out) -> inp :-> out

pattern FI :: (forall out'. Instr (ToTs inp) out') -> inp :-> out

iGenericIf :: (forall s'. Instr (ToTs a) s' -> Instr (ToTs b) s' -> Instr (ToTs c) s') -> (a :-> s) -> (b :-> s) -> c :-> s

iAnyCode :: (inp :-> out) -> Instr (ToTs inp) (ToTs out)

iNonFailingCode :: HasCallStack => (inp :-> out) -> Instr (ToTs inp) (ToTs out)

iMapAnyCode :: (forall o'. Instr (ToTs i1) o' -> Instr (ToTs i2) o') -> (i1 :-> o) -> i2 :-> o

iForceNotFail :: (i :-> o) -> i :-> o

iWithVarAnnotations :: HasCallStack => [Text] -> (inp :-> out) -> inp :-> out

Wrap Lorentz instruction with variable annotations, annots list has to be non-empty, otherwise this function raises an error.

parseLorentzValue :: forall v. (IsoValue v, SingI (ToT v), Typeable (ToT v)) => Text -> Either ParseLorentzError v

Parse textual representation of a Michelson value and turn it into corresponding Haskell value.

Note: it won't work in some complex cases, e. g. if there is a lambda which uses an instruction which depends on current contract's type. Obviously it can not work, because we don't have any information about a contract to which this value belongs (there is no such contract at all).

transformStringsLorentz :: Bool -> (MText -> MText) -> (inp :-> out) -> inp :-> out

Lorentz version of transformStrings.

transformBytesLorentz :: Bool -> (ByteString -> ByteString) -> (inp :-> out) -> inp :-> out

Lorentz version of transformBytes.

optimizeLorentz :: (inp :-> out) -> inp :-> out

optimizeLorentzWithConf :: OptimizerConf -> (inp :-> out) -> inp :-> out

class MapLorentzInstr instr where

Applicable for wrappers over Lorentz code.

Methods

mapLorentzInstr :: (forall i o. (i :-> o) -> i :-> o) -> instr -> instr

Modify all the code under given entity.

type ContractOut st = '[([Operation], st)]

type ContractCode cp st = '[(cp, st)] :-> ContractOut st

data SomeContract where

Constructors

SomeContract :: (NiceParameterFull cp, NiceStorage st) => ContractCode cp st -> SomeContract

type Lambda i o = '[i] :-> '[o]