Lorentz.Constraints.Derivative

Description

Some derivative constraints.

They are moved to separate module because they need to lie quite high in modules dependencies graph (unlike Lorentz.Constraints.Scopes).

Documentation

type NiceParameterFull cp = (Typeable cp, ParameterDeclaresEntryPoints cp)

Constraint applied to a whole parameter type.