Lorentz.Constraints.Scopes

Description

Scope-related constraints used in Lorentz.

This contains constraints from Scope modified for use in Lorentz.

Grouped constraints

type NiceParameter a = (KnownValue a, ProperParameterBetterErrors (ToT a))

Constraint applied to any part of parameter type.

Note that you don't usually apply this constraint to the whole parameter, consider using NiceParameterFull in such case.

Using this type is justified e.g. when calling another contract, there you usually supply an entrypoint argument, not the whole parameter.

type NiceStorage a = (KnownValue a, ProperStorageBetterErrors (ToT a))

type NiceConstant a = (KnownValue a, ProperConstantBetterErrors (ToT a))

type NicePackedValue a = (KnownValue a, ProperPackedValBetterErrors (ToT a))

type NiceUnpackedValue a = (KnownValue a, ProperUnpackedValBetterErrors (ToT a))

type NiceFullPackedValue a = (NicePackedValue a, NiceUnpackedValue a)

type NicePrintedValue a = (KnownValue a, ProperPrintedValBetterErrors (ToT a))

type NiceComparable a = (KnownValue a, ProperComparabilityBetterErrors (ToT a))

niceParameterEvi :: forall a. NiceParameter a :- ParameterScope (ToT a)

niceStorageEvi :: forall a. NiceStorage a :- StorageScope (ToT a)

niceConstantEvi :: forall a. NiceConstant a :- ConstantScope (ToT a)

nicePackedValueEvi :: forall a. NicePackedValue a :- PackedValScope (ToT a)

niceUnpackedValueEvi :: forall a. NiceUnpackedValue a :- UnpackedValScope (ToT a)

nicePrintedValueEvi :: forall a. NicePrintedValue a :- PrintedValScope (ToT a)

Individual constraints (internals)

class (IsoValue a, HasNoNestedBigMaps (ToT a)) => CanHaveBigMap a

class (IsoValue a, Typeable (ToT a), SingI (ToT a)) => KnownValue a

Gathers constraints, commonly required for values.

class (IsoValue a, Typeable (ToCT a), SingI (ToCT a)) => KnownCValue a

class (IsoValue a, ForbidOp (ToT a)) => NoOperation a

Ensure given type does not contain "operation".

class (IsoValue a, ForbidContract (ToT a)) => NoContractType a

class (IsoValue a, ForbidBigMap (ToT a)) => NoBigMap a

Re-exports

withDict :: HasDict c e => e -> (c -> r) -> r

From a Dict, takes a value in an environment where the instance witnessed by the Dict is in scope, and evaluates it.

Essentially a deconstruction of a Dict into its continuation-style form.

Can also be used to deconstruct an entailment, a :- b, using a context a.

withDict :: Dict c -> (c => r) -> r
withDict :: a => (a :- c) -> (c => r) -> r