Lorentz.ContractRegistry

Description

This module contains various datatypes and functions which are common for contract registry packages (e.g. morley-ledgers and morley-multisig).

Registry types

data ContractInfo

Constructors

(NiceParameterFull cp, NiceStorage st) => ContractInfo

Fields

ciContract :: ContractCode cp st

ciIsDocumented :: Bool

ciStorageParser :: Maybe (Parser st)

Specifies how to parse initial storage value.

Normally you pass some user data and call a function that constructs storage from that data.

If storage is simple and can be easilly constructed manually, you can use Nothing.

ciCompilationOptions :: CompilationOptions

How to compile this contract.

ciStorageNotes :: Notes (ToT st)

A temporary approach to add annotations to storage. TODO [#20]: invent something better.

newtype ContractRegistry

Constructors

ContractRegistry

Fields

unContractRegistry :: Map Text ContractInfo

(?::) :: Text -> a -> (Text, a)

Things to do in main

data CmdLnArgs

ContractRegistry actions parsed from CLI.

Constructors

List

Print Text (Maybe FilePath) Bool Bool

Document Text (Maybe FilePath) DGitRevision

PrintStorage SomeNiceStorage Bool

argParser :: ContractRegistry -> DGitRevision -> Parser CmdLnArgs

runContractRegistry :: ContractRegistry -> CmdLnArgs -> IO ()

Run an action operating with ContractRegistry.

Building blocks

getContract :: Text -> ContractRegistry -> Either String ContractInfo

printContractFromRegistryDoc :: Text -> ContractRegistry -> DGitRevision -> Maybe FilePath -> IO ()