Lorentz.Empty

Description

Support for uninhabited type.

Currently they are not supported my Michelson, so we provide a sort of replacement.

This module should be removed once the proposal is implemented: https://gitlab.com/tezos/tezos/issues/662

Documentation

data Empty

Replacement for uninhabited type.

absurd_ :: (Empty ': s) :-> s'

Witness of that this code is unreachable.

Orphan instances

CustomErrorHasDoc "emptySupplied"

Instance details

Methods

customErrDocMdCause :: Markdown

customErrDocMdCauseInEntrypoint :: Markdown

customErrClass :: ErrorClass

customErrArgumentSemantics :: Maybe Markdown

Buildable (CustomError "emptySupplied")

Instance details

Methods

build :: CustomError "emptySupplied" -> Builder