Lorentz.EntryPoints.Core

Description

Primitives supplying entrypoints declarations and lookup.

Documentation

type family CanHaveEntryPoints (p :: Type) :: Bool where ...

Used to understand whether a type can potentially declare any entrypoints.

Equations

CanHaveEntryPoints (ShouldHaveEntryPoints _) = True

CanHaveEntryPoints p = CanHaveEntryPointsT (ToT p)

class EntryPointsDerivation deriv cp where

Defines a generalized way to declare entrypoints for various parameter types.

When defining instances of this typeclass, set concrete deriv argument and leave variable cp argument. Also keep in mind, that in presence of explicit default entrypoint, all other Or arms should be callable, though you can put this burden on user if very necessary.

Methods of this typeclass aim to better type-safety when making up an implementation and they may be not too convenient to use; users should exploit their counterparts.

Associated Types

type EpdAllEntryPoints deriv cp :: [(Symbol, Type)]

Name and argument of each entrypoint. This may include intermediate ones, even root if necessary.

Touching this type family is costly (O(N^2)), don't use it often.

Note [order of entrypoints children]: If this contains entrypoints referring to indermediate nodes (not leaves) in or tree, then each such entrypoint should be mentioned eariler than all of its children.

type EpdLookupEntryPoint deriv cp :: Symbol -> Exp (Maybe Type)

Get entrypoint argument by name.

Methods

epdNotes :: Notes (ToT cp)

Construct parameter annotations corresponding to expected entrypoints set.

This method is implementation detail, for actual notes construction use parameterEntryPointsToNotes.

TODO [#35]: Should also return field annotation

epdCall :: ParameterScope (ToT cp) => Label name -> EpConstructionRes (ToT cp) (Eval (EpdLookupEntryPoint deriv cp name))

Construct entrypoint caller.

This does not treat calls to default entrypoint in a special way.

This method is implementation detail, for actual entrypoint lookup use parameterEntryPointCall.

epdDescs :: Rec EpCallingDesc (EpdAllEntryPoints deriv cp)

Description of how each of the entrypoints is constructed.

data EpConstructionRes (param :: T) (marg :: Maybe Type) where

Result of entrypoint lookup at term level.

Constructors

EpConstructed :: ParameterScope (ToT arg) => EpLiftSequence (ToT arg) param -> EpConstructionRes param (Just arg)

EpConstructionFailed :: EpConstructionRes param Nothing

data EpCallingDesc (info :: (Symbol, Type)) where

How one of the entrypoints is called.

Type arguments are name of the constructor which eventually gave name to the entrypoint and this entrypoint's argument.

Constructors

EpCallingDesc

Fields

:: { epcdArg :: Proxy (arg :: Type)

Entrypoint argument type.

, epcdEntrypoint :: EpName

Name of assigned entrypoint.

, epcdSteps :: [EpCallingStep]

If we emulated entrypoints calling via just wrapping an argument into constructors until getting the full parameter, how would it look like. Steps are enlisted in reversed order - top-level constructors go last.

} -> EpCallingDesc '(name, arg)

data EpCallingStep

Constructors

EpsWrapIn Text
Wrap into constructor with given name.

type RequireAllUniqueEntryPoints cp = RequireAllUniqueEntryPoints' (ParameterEntryPointsDerivation cp) cp

Ensure that all declared entrypoints are unique.

class (EntryPointsDerivation (ParameterEntryPointsDerivation cp) cp, RequireAllUniqueEntryPoints cp) => ParameterHasEntryPoints cp

Which entrypoints given parameter declares.

Note that usually this function should not be used as constraint, use ParameterDeclaresEntryPoints for this purpose.

Associated Types

type ParameterEntryPointsDerivation cp :: Type

type ParameterDeclaresEntryPoints cp = (If (CanHaveEntryPoints cp) (ParameterHasEntryPoints cp) (() :: Constraint), NiceParameter cp, EntryPointsDerivation (GetParameterEpDerivation cp) cp)

Parameter declares some entrypoints.

This is a version of ParameterHasEntryPoints which we actually use in constraints. When given type is a sum type or newtype, we refer to ParameterHasEntryPoints instance, otherwise this instance is not necessary.

type GetParameterEpDerivation cp = If (CanHaveEntryPoints cp) (ParameterEntryPointsDerivation cp) EpdNone

Version of ParameterEntryPointsDerivation which we actually use in function signatures. When given type is sum type or newtype, we refer to ParameterEntryPointsDerivation, otherwise we suppose that no entrypoints are declared.

pepNotes :: forall cp. ParameterDeclaresEntryPoints cp => Notes (ToT cp)

Version of epdNotes which we actually use in code. It hides derivations stuff inside, and treats primitive types specially like GetParameterEpDerivation does.

pepCall :: forall cp name. (ParameterDeclaresEntryPoints cp, ParameterScope (ToT cp)) => Label name -> EpConstructionRes (ToT cp) (Eval (LookupParameterEntryPoint cp name))

Version of epdCall which we actually use in code. It hides derivations stuff inside, and treats primitive types specially like GetParameterEpDerivation does.

pepDescs :: forall cp. ParameterDeclaresEntryPoints cp => Rec EpCallingDesc (AllParameterEntryPoints cp)

Version of epdDescs which we actually use in code. It hides derivations stuff inside, and treats primitive types specially like GetParameterEpDerivation does.

pepDescsWithDef :: forall cp. ParameterDeclaresEntryPoints cp => [Some1 EpCallingDesc]

Descriptions of how each of the entrypoints is constructed.

Similar to pepDescs, but includes default entrypoint disregard whether it is explicit or not, while pepDescs includes it only if it is explicit. Also this returns list, not Rec, for simplicity.

Note that [order of entrypoints children] property still holds here.

type family AllParameterEntryPoints (cp :: Type) :: [(Symbol, Type)] where ...

Get all entrypoints declared for parameter.

Equations

AllParameterEntryPoints cp = EpdAllEntryPoints (GetParameterEpDerivation cp) cp

type family LookupParameterEntryPoint (cp :: Type) :: Symbol -> Exp (Maybe Type) where ...

Lookup for entrypoint type by name.

Does not treat default entrypoints in a special way.

Equations

LookupParameterEntryPoint cp = EpdLookupEntryPoint (GetParameterEpDerivation cp) cp

parameterEntryPointsToNotes :: forall cp. (Typeable cp, ParameterDeclaresEntryPoints cp) => ParamNotes (ToT cp)

Derive annotations for given parameter.

type GetEntryPointArg cp name = Eval (LiftM2 FromMaybe (TError ((Text "Entrypoint not found: " :<>: ShowType name) :$$: ((Text "In contract parameter `" :<>: ShowType cp) :<>: Text "`"))) (LookupParameterEntryPoint cp name))

Get type of entrypoint with given name, fail if not found.

parameterEntryPointCall :: forall cp name. ParameterDeclaresEntryPoints cp => Label name -> EntryPointCall cp (GetEntryPointArg cp name)

Prepare call to given entrypoint.

This does not treat calls to default entrypoint in a special way. To call default entrypoint properly use parameterEntryPointCallDefault.

type GetDefaultEntryPointArg cp = Eval (LiftM2 FromMaybe (Pure cp) (LookupParameterEntryPoint cp DefaultEpName))

Get type of entrypoint with given name, fail if not found.

parameterEntryPointCallDefault :: forall cp. ParameterDeclaresEntryPoints cp => EntryPointCall cp (GetDefaultEntryPointArg cp)

Call the default entrypoint.

type ForbidExplicitDefaultEntryPoint cp = Eval (LiftM3 UnMaybe (Pure (Pure (() :: Constraint))) (TError (Text "Parameter used here must have no explicit \"default\" entrypoint" :$$: ((Text "In parameter type `" :<>: ShowType cp) :<>: Text "`"))) (LookupParameterEntryPoint cp DefaultEpName))

Ensure that there is no explicit "default" entrypoint.

type NoExplicitDefaultEntryPoint cp = Eval (LookupParameterEntryPoint cp DefaultEpName) ~ Nothing

Similar to ForbidExplicitDefaultEntryPoint, but in a version which the compiler can work with (and which produces errors confusing for users :/)

sepcCallRootChecked :: forall cp. (NiceParameter cp, ForbidExplicitDefaultEntryPoint cp) => SomeEntryPointCall cp

Call root entrypoint safely.

data EntryPointRef (mname :: Maybe Symbol) where

Which entrypoint to call.

We intentionally distinguish default and non-default cases because this makes API more details-agnostic.

Constructors

CallDefault :: EntryPointRef Nothing
Call the default entrypoint, or root if no explicit default is assigned.

Call :: NiceEntryPointName name => EntryPointRef (Just name)
Call the given entrypoint; calling default is not treated specially. You have to provide entrypoint name via passing it as type argument.

Unfortunatelly, here we cannot accept a label because in most cases our entrypoints begin from capital letter (being derived from constructor name), while labels must start from a lower-case letter, and there is no way to make a conversion at type-level.

type NiceEntryPointName name = (KnownSymbol name, ForbidDefaultName name)

Constraint on type-level entrypoint name specifier.

eprName :: forall mname. EntryPointRef mname -> EpName

type family GetEntryPointArgCustom cp mname :: Type where ...

Universal entrypoint lookup.

Equations

GetEntryPointArgCustom cp Nothing = GetDefaultEntryPointArg cp

GetEntryPointArgCustom cp (Just name) = GetEntryPointArg cp name

newtype TrustEpName

This wrapper allows to pass untyped EpName and bypass checking that entrypoint with given name and type exists.

Constructors

TrustEpName EpName

class HasEntryPointArg cp name arg where

When we call a Lorentz contract we should pass entrypoint name and corresponding argument. Ideally we want to statically check that parameter has entrypoint with given name and argument. Constraint defined by this type class holds for contract with parameter cp that have entrypoint matching name with type arg.

In order to check this property statically, we need to know entrypoint name in compile time, EntryPointRef type serves this purpose. If entrypoint name is not known, one can use TrustEpName wrapper to take responsibility for presence of this entrypoint.

If you want to call a function which has this constraint, you have two options: 1. Pass contract parameter cp using type application, pass EntryPointRef as a value and pass entrypoint argument. Type system will check that cp has an entrypoint with given reference and type. 2. Pass EpName wrapped into TrustEpName and entrypoint argument. In this case passing contract parameter is not necessary, you do not even have to know it.

Methods

useHasEntryPointArg :: name -> (Dict (ParameterScope (ToT arg)), EpName)

Data returned by this method may look somewhat arbitrary. EpName is obviously needed because name can be EntryPointRef or TrustEpName. Dict is returned because in EntryPointRef case we get this evidence for free and don't want to use it. We seem to always need it anyway.

type HasDefEntryPointArg cp defEpName defArg = (defEpName ~ EntryPointRef Nothing, HasEntryPointArg cp defEpName defArg)

HasEntryPointArg constraint specialized to default entrypoint.

type HasEntryPointOfType param con exp = (GetEntryPointArgCustom param (Just con) ~ exp, ParameterDeclaresEntryPoints param)

Checks that the given parameter consists of some specific entrypoint. Similar as HasEntryPointArg but ensures that the argument matches the following datatype.

type family ParameterContainsEntryPoints param (fields :: [NamedEp]) :: Constraint where ...

Check that the given entrypoint has some fields inside. This interface allows for an abstraction of contract parameter so that it requires some *minimal* specification, but not a concrete one.

Equations

ParameterContainsEntryPoints _ '[] = ()

ParameterContainsEntryPoints param ((n :> ty) ': rest) = (HasEntryPointOfType param n ty, ParameterContainsEntryPoints param rest)

parameterEntryPointCallCustom :: forall cp mname. ParameterDeclaresEntryPoints cp => EntryPointRef mname -> EntryPointCall cp (GetEntryPointArgCustom cp mname)

Universal entrypoint calling.

data EpdNone

No entrypoints declared, parameter type will serve as argument type of the only existing entrypoint (default one).

type (:>) n ty = NamedEp n ty infixr 0

Internals

type RequireAllUniqueEntryPoints' deriv cp = RequireAllUnique "entrypoint name" (Eval (Map Fst $ EpdAllEntryPoints deriv cp))