Lorentz.EntryPoints.Doc

Description

Utilities for declaring and documenting entry points.

Documentation

data DEntryPoint (kind :: Type)

Gathers information about single entrypoint.

We assume that entry points might be of different kinds, which is designated by phantom type parameter. For instance, you may want to have several groups of entry points corresponding to various parts of a contract - specifying different kind type argument for each of those groups will allow you defining different DocItem instances with appropriate custom descriptions for them.

Constructors

DEntryPoint

Fields

depName :: Text

depSub :: SubDoc

data DEntryPointReference

Constructors

DEntryPointReference Text Anchor

class EntryArrow kind name body where

Provides arror for convenient entrypoint documentation

Methods

(#->) :: (Label name, Proxy kind) -> body -> body

Lift entrypoint implementation.

Entrypoint names should go with "e" prefix.

data PlainEntryPointsKind

Default value for DEntryPoint type argument.

diEntryPointToMarkdown :: HeaderLevel -> DEntryPoint level -> Markdown

Default implementation of docItemToMarkdown for entry points.

data DEntryPointArg

Describes argument of an entrypoint.

Constructors

DEntryPointArg

Fields

epaArg :: Maybe DType

Argument of the entrypoint. Pass Nothing if no argument is required.

epaBuilding :: [ParamBuildingStep]

Describes a way to lift an entrypoint argument into full parameter which can be passed to the contract.

Steps are supposed to be applied in the order opposite to one in which they are given. E.g. suppose that an entrypoint is called as Run (Service1 arg); then the first step (actual last) should describe wrapping into Run constructor, and the second step (actual first) should be about wrapping into Service1 constructor.

epaType :: Type

Untyped representation of entrypoint, used for printing its michelson type representation.

data DType where

Doc element with description of a type.

Constructors

DType :: forall a. TypeHasDoc a => Proxy a -> DType

class KnownSymbol con => DeriveCtorFieldDoc con (cf :: CtorField) where

Pick a type documentation from CtorField.

Methods

deriveCtorFieldDoc :: DEntryPointArg

newtype ParamBuilder

When describing the way of parameter construction - piece of incremental builder for this description.

Constructors

ParamBuilder

Fields

unParamBuilder :: Markdown -> Markdown

Argument stands for previously constructed parameter piece, and returned value - a piece constructed after our step.

data ParamBuildingDesc

Constructors

ParamBuildingDesc

Fields

pbdEnglish :: Markdown

Plain english description of this step.

pbdHaskell :: ParamBuilder

How to construct parameter in Haskell code.

pbdMichelson :: ParamBuilder

How to construct parameter working on raw Michelson.

data ParamBuildingStep

Describes a parameter building step.

This can be wrapping into (Haskell) constructor, or a more complex transformation.

Constructors

PbsWrapIn Text ParamBuildingDesc
Wraps something into constructor with given name. Constructor should be the one which corresponds to an entrypoint defined via field annotation, for more complex cases use PbsCustom.

PbsCallEntrypoint EpName
Directly call an entrypoint marked with a field annotation.

PbsCustom ParamBuildingDesc
Other action.

PbsUncallable [ParamBuildingStep]
This entrypoint cannot be called, which is possible when an explicit default entrypoint is present. This is not a true entrypoint but just some intermediate node in or tree and neither it nor any of its parents are marked with a field annotation.

It contains dummy ParamBuildingSteps which were assigned before entrypoints were taken into account.

mkPbsWrapIn :: Text -> ParamBuilder -> ParamBuildingStep

Make a ParamBuildingStep that tells about wrapping an argument into a constructor with given name and uses given ParamBuilder as description of Michelson part.

clarifyParamBuildingSteps :: ParamBuildingStep -> (inp :-> out) -> inp :-> out

Go over contract code and update every occurrence of DEntryPointArg documentation item, adding the given step to its "how to build parameter" description.

constructDEpArg :: forall arg. (TypeHasDoc arg, HasTypeAnn arg, KnownValue arg) => DEntryPointArg

emptyDEpArg :: DEntryPointArg

mkUType :: SingI x => Notes x -> Type

mkDEpUType :: forall t. (KnownValue t, HasTypeAnn t) => Type

mkDEntryPointArgSimple :: forall t. (KnownValue t, HasTypeAnn t, TypeHasDoc t) => DEntryPointArg

type DocumentEntryPoints kind a = (Generic a, GDocumentEntryPoints kind (Rep a))

Constraint for documentEntryPoints.

documentEntryPoint :: forall kind epName param s out. (KnownSymbol epName, DocItem (DEntryPoint kind), TypeHasDoc param, HasTypeAnn param, KnownValue param) => ((param & s) :-> out) -> (param & s) :-> out

Wrapper for documenting single entrypoint which parameter isn't going to be unwrapped from some datatype.

entryCase unwraps a datatype, however, sometimes we want to have entrypoint parameter to be not wrapped into some datatype.

entryCase :: forall dt entryPointKind out inp clauses. (CaseTC dt out inp clauses, DocumentEntryPoints entryPointKind dt) => Proxy entryPointKind -> IsoRecTuple clauses -> (dt & inp) :-> out

Version of entryCase_ for tuples.

entryCase_ :: forall dt entryPointKind out inp. (InstrCaseC dt, RMap (CaseClauses dt), DocumentEntryPoints entryPointKind dt) => Proxy entryPointKind -> Rec (CaseClauseL inp out) (CaseClauses dt) -> (dt & inp) :-> out

Like case_, to be used for pattern-matching on a parameter or its part.

Modifies documentation accordingly. Including description of entrypoints' arguments, thus for them you will need to supply TypeHasDoc instance.

finalizeParamCallingDoc :: forall cp inp out. (NiceParameterFull cp, RequireSumType cp, HasCallStack) => ((cp ': inp) :-> out) -> (cp ': inp) :-> out

Modify param building steps with respect to entrypoints that given parameter declares.

Each contract with entrypoints should eventually call this function, otherwise, in case if contract uses built-in entrypoints feature, the resulting parameter building steps in the generated documentation will not consider entrypoints and thus may be incorrect.

Calling this twice over the same code is also prohibited.

areFinalizedParamBuildingSteps :: [ParamBuildingStep] -> Bool

Whether finalizeParamCallingDoc has already been applied to these steps.

entryCaseSimple_ :: forall cp out inp. (InstrCaseC cp, RMap (CaseClauses cp), DocumentEntryPoints PlainEntryPointsKind cp, NiceParameterFull cp, RequireFlatParamEps cp) => Rec (CaseClauseL inp out) (CaseClauses cp) -> (cp & inp) :-> out

entryCaseSimple :: forall cp out inp clauses. (CaseTC cp out inp clauses, DocumentEntryPoints PlainEntryPointsKind cp, NiceParameterFull cp, RequireFlatParamEps cp) => IsoRecTuple clauses -> (cp & inp) :-> out

Version of entryCase for contracts with flat parameter, use it when you need only one entryCase all over the contract implementation.

This method calls finalizeParamCallingDoc inside.

type family RequireFlatParamEps cp :: Constraint where ...

Equations

RequireFlatParamEps cp = (RequireFlatEpDerivation cp (GetParameterEpDerivation cp), RequireSumType cp)

type family RequireFlatEpDerivation cp deriv :: Constraint where ...

Equations

RequireFlatEpDerivation _ EpdNone = ()

RequireFlatEpDerivation _ EpdPlain = ()

RequireFlatEpDerivation cp deriv = TypeError ((Text "Parameter is not flat" :$$: ((Text "For parameter `" :<>: ShowType cp) :<>: Text "`")) :$$: ((Text "With entrypoints derivation way `" :<>: ShowType deriv) :<>: Text "`"))