Lorentz.EntryPoints.Helpers

Documentation

ctorNameToAnn :: forall ctor. (KnownSymbol ctor, HasCallStack) => FieldAnn

ctorNameToEp :: forall ctor. (KnownSymbol ctor, HasCallStack) => EpName

type family CanHaveEntryPoints (p :: Type) :: Bool where ...

Used to understand whether a type can potentially declare any entrypoints.

Equations

CanHaveEntryPoints (ShouldHaveEntryPoints _) = True

CanHaveEntryPoints p = CanHaveEntryPointsT (ToT p)

newtype ShouldHaveEntryPoints a

A special type which wraps over a primitive type and states that it has entrypoints (one).

Assuming that any type can have entrypoints makes use of Lorentz entrypoints too annoying, so for declaring entrypoints for not sum types we require an explicit wrapper.

Constructors

ShouldHaveEntryPoints

Fields

unHasEntryPoints :: a

type family RequireSumType (a :: Type) :: Constraint where ...

Ensure that given type is a sum type.

This helps to prevent attempts to apply a function to, for instance, a pair.

Equations

RequireSumType a = If (CanHaveEntryPoints a) (() :: Constraint) (TypeError (Text "Expected Michelson sum type" :$$: ((Text "In type `" :<>: ShowType a) :<>: Text "`")))