Lorentz.Errors

Haskell to Value conversion

class (Typeable e, ErrorHasDoc e) => IsError e where

Haskell type representing error.

Methods

errorToVal :: e -> (forall t. ErrorScope t => Value t -> r) -> r

Converts a Haskell error into Value representation.

errorFromVal :: (Typeable t, SingI t) => Value t -> Either Text e

Converts a Value into Haskell error.

isoErrorToVal :: (KnownError e, IsoValue e) => e -> (forall t. ErrorScope t => Value t -> r) -> r

Implementation of errorToVal via IsoValue.

isoErrorFromVal :: (Typeable t, Typeable (ToT e), IsoValue e) => Value t -> Either Text e

Implementation of errorFromVal via IsoValue.

class Typeable e => ErrorHasDoc (e :: Type) where

Minimal complete definition

errorDocName, errorDocMdCause, errorDocHaskellRep, errorDocDependencies

Associated Types

type ErrorRequirements e :: Constraint

Constraints which we require in a particular instance. You are not oblidged to often instantiate this correctly, it is only useful for some utilities.

Methods

errorDocName :: Text

Name of error as it appears in the corresponding section title.

errorDocMdCause :: Markdown

What should happen for this error to be raised.

errorDocMdCauseInEntrypoint :: Markdown

Brief version of errorDocMdCause.

This will appear along with the error when mentioned in entrypoint description. By default, the first sentence of the full description is used.

errorDocHaskellRep :: Markdown

How this error is represented in Haskell.

errorDocClass :: ErrorClass

Error class.

errorDocDependencies :: [SomeDocDefinitionItem]

Which definitions documentation for this error mentions.

errorDocRequirements :: Dict (ErrorRequirements e)

Captured constraints which we require in a particular instance. This is a way to encode a bidirectional instance in the nowaday Haskell, for class MyConstraint => ErrorHasDoc MyType instance it lets deducing MyConstraint by ErrorHasDoc MyType.

You are not oblidged to always instantiate, it is only useful for some utilities which otherwise would not compile.

errorDocRequirements :: ErrorRequirements e => Dict (ErrorRequirements e)

Captured constraints which we require in a particular instance. This is a way to encode a bidirectional instance in the nowaday Haskell, for class MyConstraint => ErrorHasDoc MyType instance it lets deducing MyConstraint by ErrorHasDoc MyType.

You are not oblidged to always instantiate, it is only useful for some utilities which otherwise would not compile.

typeDocMdDescriptionReferToError :: forall e. IsError e => Markdown

Implementation of typeDocMdDescription (of TypeHasDoc typeclass) for Haskell types which sole purpose is to be error.

customErrorDocHaskellRepGeneral :: (SingI (ToT (ErrorArg tag)), IsError (CustomError tag), TypeHasDoc (ErrorArg tag), CustomErrorHasDoc tag) => Text -> Proxy tag -> Markdown

Description of error representation in Haskell.

data UnspecifiedError

Use this type as replacement for () when you really want to leave error cause unspecified.

Constructors

UnspecifiedError

General instructions

failUsing :: forall e s t. IsError e => e -> s :-> t

Fail with the given Haskell value.

failUnexpected :: MText -> s :-> t

Fail, providing a reference to the place in the code where this function is called.

Like error in Haskell code, this instruction is for internal errors only.

Custom errors

type family ErrorArg (tag :: Symbol) :: Type

Declares a custom error, defining error name - error argument relation.

If your error is supposed to carry no argument, then provide ().

Note that this relation is defined globally rather than on per-contract basis, so define errors accordingly. If your error has argument specific to your contract, call it such that error name reflects its belonging to this contract.

This is the basic [error format].

data CustomError (tag :: Symbol)

Material custom error.

Use this in pattern matches against error (e.g. in tests).

Constructors

CustomError

Fields

ceTag :: Label tag

ceArg :: ErrorArg tag

failCustom :: forall tag err s any. (err ~ ErrorArg tag, CustomErrorHasDoc tag, KnownError err) => Label tag -> (err ': s) :-> any

Fail with given custom error.

type RequireNoArgError tag msg = (TypeErrorUnless (ErrorArg tag == ()) msg, msg ~ (Text "Expected no-arg error, but given error requires argument of type " :<>: ShowType (ErrorArg tag)))

failCustom_ :: forall tag s any notVoidErrorMsg. (RequireNoArgError tag notVoidErrorMsg, CustomErrorHasDoc tag) => Label tag -> s :-> any

Specialization of failCustom for no-arg errors.

Documentation

data ErrorClass

Error class on how the error should be handled by the client.

Constructors

ErrClassActionException
Normal expected error. Examples: "insufficient balance", "wallet does not exist".

ErrClassBadArgument
Invalid argument passed to entrypoint. Examples: your entrypoint accepts an enum represented as nat, and unknown value is provided. This includes more complex cases which involve multiple entrypoints. E.g. API provides iterator interface, middleware should care about using it hiding complex details and exposing a simpler API to user; then an attempt to request non-existing element would also correspond to an error from this class.

ErrClassContractInternal
Unexpected error. Most likely it means that there is a bug in the contract or the contract has been deployed incorrectly.

ErrClassUnknown
It's possible to leave error class unspecified.

class (KnownSymbol tag, TypeHasDoc (ErrorArg tag), IsError (CustomError tag)) => CustomErrorHasDoc tag where

Minimal complete definition

customErrDocMdCause, customErrClass

Methods

customErrDocMdCause :: Markdown

What should happen for this error to be raised.

customErrDocMdCauseInEntrypoint :: Markdown

Brief version of customErrDocMdCause. This will appear along with the error when mentioned in entrypoint description.

By default, the first sentence of the full description is used.

customErrClass :: ErrorClass

Error class.

By default this returns "unknown error" class; though you should provide explicit implementation in order to avoid a warning.

customErrArgumentSemantics :: Maybe Markdown

Clarification of error argument meaning.

Provide when it's not obvious, e.g. argument is not named with :!.

NOTE: This should not be an entire sentence, rather just the semantic backbone.

Bad: * Error argument stands for the previous value of approval.

Good: * the previous value of approval * pair, first argument of which is one thing, and the second is another

data DError where

Mentions that contract uses given error.

Constructors

DError :: ErrorHasDoc e => Proxy e -> DError

data DThrows where

Documentation for custom errors.

Mentions that entrypoint throws given error.

Constructors

DThrows :: ErrorHasDoc e => Proxy e -> DThrows

Old interface (DEPRECATED)

customErrorToVal :: (LooseSumC e, HasCallStack) => e -> (forall t. ErrorScope t => Value t -> r) -> r

Deprecated: Datatype error declarations has been deprecated

Implementation of errorToVal for custom errors.

customErrorFromVal :: forall t e. (SingI t, LooseSumC e) => Value t -> Either Text e

Deprecated: Datatype error declarations has been deprecated

Implementation of errorFromVal for custom errors.

This function is deprecated.

failUsingArg :: forall err name fieldTy s s'. FailUsingArg err name fieldTy s s'

Deprecated: Datatype error declarations has been deprecated

Fail with given error, picking argument for error from the top of the stack.

If your error constructor does not carry an argument, use failUsing function instead. Consider the following practice: once error datatype for your contract is defined, create a specialization of this function to the error type.

This function is deprecated.

type FailUsingArg e name fieldTy s s' = (KnownSymbol name, IsError e, IsoValue fieldTy, CtorHasOnlyField name e fieldTy, Each [Typeable, SingI] '[ToT fieldTy], HasCallStack) => Label name -> (fieldTy ': s) :-> s'

Signature of userFailWith.

type family CustomErrorNoIsoValue a where ...

Prompt an error message saying that IsoValue is not applicable for this type.

Equations

CustomErrorNoIsoValue a = TypeError ((Text "No IsoValue instance for " :<>: ShowType a) :$$: Text "It has custom error representation")

deriveCustomError :: Name -> Q [Dec]

Deprecated: Datatype error declarations has been deprecated

Derive IsError instance for given type.

This will also forbid deriving IsoValue instance for that type to avoid having multiple different Michelson representations.

Internals

errorTagToText :: forall tag. KnownSymbol tag => Text

errorTagToMText :: Label tag -> MText

Demote error tag to term level.