Lorentz.Errors.Numeric.Doc

Description

Autodoc for numeric errors.

Documentation

data DDescribeErrorTagMap

Adds a section which explains error tag mapping.

Constructors

DDescribeErrorTagMap

Fields

detmSrcLoc :: Text

Describes where the error tag map is defined in Haskell code.

applyErrorTagToErrorsDoc :: HasCallStack => ErrorTagMap -> (inp :-> out) -> inp :-> out

Modify documentation generated for given code so that all CustomError mention not their textual error tag rather respective numeric one from the given map.

If some documented error is not present in the map, it remains unmodified. This function may fail with error if contract uses some uncommon errors, see applyErrorTagToErrorsDocWith for details.

applyErrorTagToErrorsDocWith :: HasCallStack => [NumericErrorDocHandler] -> ErrorTagMap -> (inp :-> out) -> inp :-> out

Extended version of applyErrorTagToErrorsDoc which accepts error handlers.

In most cases that function should be enough for your purposes, but it uses a fixed set of base handlers which may be not enough in case when you define your own errors. In this case define and pass all the necessary handlers to this function.

It fails with error if some of the errors used in the contract cannot be handled with given handlers.

data NumericErrorDocHandler

Handler which changes documentation for one particular error type.

data NumericErrorDocHandlerError

Errors for NumericErrorDocHandler

customErrorDocHandler :: NumericErrorDocHandler

Handler for all CustomErrors.

voidResultDocHandler :: NumericErrorDocHandler

Handler for VoidResult.

baseErrorDocHandlers :: [NumericErrorDocHandler]

Handlers for most common errors defined in Lorentz.

Internals

data NumericErrorWrapper (numTag :: Nat) (err :: Type)

Some error with a numeric tag attached.