Lorentz.Polymorphic

Description

Type families from Polymorphic lifted to Haskell types.

Documentation

class (MemOp (ToT c), ToT (MemOpKeyHs c) ~ Tc (MemOpKey (ToT c))) => MemOpHs c

Lifted MemOpKey.

Associated Types

type MemOpKeyHs c :: Type

class (MapOp (ToT c), ToT (MapOpInpHs c) ~ MapOpInp (ToT c), ToT (MapOpResHs c ()) ~ MapOpRes (ToT c) (ToT ())) => MapOpHs c

Lifted MapOp.

Associated Types

type MapOpInpHs c :: Type

type MapOpResHs c :: Type -> Type

class (IterOp (ToT c), ToT (IterOpElHs c) ~ IterOpEl (ToT c)) => IterOpHs c

Lifted IterOp.

Associated Types

type IterOpElHs c :: Type

class SizeOp (ToT c) => SizeOpHs c

Lifted SizeOp.

This could be just a constraint alias, but to avoid T types appearance in error messages we make a full type class with concrete instances.

class (UpdOp (ToT c), ToT (UpdOpKeyHs c) ~ Tc (UpdOpKey (ToT c)), ToT (UpdOpParamsHs c) ~ UpdOpParams (ToT c)) => UpdOpHs c

Lifted UpdOp.

Associated Types

type UpdOpKeyHs c :: Type

type UpdOpParamsHs c :: Type

class (GetOp (ToT c), ToT (GetOpKeyHs c) ~ Tc (GetOpKey (ToT c)), ToT (GetOpValHs c) ~ GetOpVal (ToT c)) => GetOpHs c

Lifted GetOp.

Associated Types

type GetOpKeyHs c :: Type

type GetOpValHs c :: Type

class ConcatOp (ToT c) => ConcatOpHs c

Lifted ConcatOp.

class SliceOp (ToT c) => SliceOpHs c

Lifted SliceOp.

class (EDivOp (ToCT n) (ToCT m), IsComparable n, IsComparable m, ToT (EDivOpResHs n m) ~ Tc (EDivOpRes (ToCT n) (ToCT m)), ToT (EModOpResHs n m) ~ Tc (EModOpRes (ToCT n) (ToCT m))) => EDivOpHs n m

Lifted EDivOp.

Associated Types

type EDivOpResHs n m :: Type

type EModOpResHs n m :: Type

type family IsoMapOpRes c b where ...

A useful property which holds for reasonable MapOp instances.

It's a separate thing from MapOpHs because it mentions b type parameter.

Equations

IsoMapOpRes c b = ToT (MapOpResHs c b) ~ MapOpRes (ToT c) (ToT b)