Lorentz.Run

Documentation

data CompilationOptions

Constructors

CompilationOptions

Fields

coDisableInitialCast :: Bool

Flag which defines, whether compiled Michelson contract will have CAST (which drops parameter annotations) as a first instruction. Note, that when flag is false, there still can be no CAST (in case when parameter type has no annotations).

defaultCompilationOptions :: CompilationOptions

compileLorentz :: (inp :-> out) -> Instr (ToTs inp) (ToTs out)

For use outside of Lorentz.

compileLorentzContract :: forall cp st. (NiceParameterFull cp, NiceStorage st) => ContractCode cp st -> FullContract (ToT cp) (ToT st)

Version of compileLorentz specialized to instruction corresponding to contract code.

compileLorentzContractWithOptions :: forall cp st. (NiceParameterFull cp, NiceStorage st) => CompilationOptions -> ContractCode cp st -> FullContract (ToT cp) (ToT st)

Version on compileLorentzContract which accepts CompilationOptions.

Note that compiled contract can be ill-typed in terms of Michelson code when some of the compilation options are used (e.g. when coDoInitialCast is False, resulted contract can be ill-typed). However, compilation with defaultCompilationOptions should be valid.

interpretLorentzInstr :: (IsoValuesStack inp, IsoValuesStack out) => ContractEnv -> (inp :-> out) -> Rec Identity inp -> Either MichelsonFailed (Rec Identity out)

Interpret a Lorentz instruction, for test purposes.

interpretLorentzLambda :: (IsoValue inp, IsoValue out) => ContractEnv -> Lambda inp out -> inp -> Either MichelsonFailed out

Like interpretLorentzInstr, but works on lambda rather than arbitrary instruction.

analyzeLorentz :: (inp :-> out) -> AnalyzerRes

Lorentz version of analyzer.