Lorentz.Test.Consumer

Description

Contract which remembers all parameters it has been called with.

Useful to save return values of View entry points.

Documentation

contractConsumer :: ContractCode cp [cp]

Remembers parameters it was called with, last goes first.