Lorentz.TestScenario

Description

Deprecated: Use morley-nettest instead

Documentation

type TestScenario param = [(Address, param)]

Type that represents test scenario for Lorentz contract. Simply put, it is sequence of pairs (sender, parameter). Using this sequence we can perform transfers to the desired contract.

showTestScenario :: (Buildable param, NicePrintedValue param) => TestScenario param -> Text

Function to get textual representation of TestScenario, each Parameter is printed as a raw Michelson value. This representation can later be used in order to run test scenario on real network.

The format for a single contract call is the following: # `printed Lorentz parameter` (actually comment) `sender address` `printed raw Michelson parameter`