Lorentz.UParam

Documentation

newtype UParam (entries :: [EntryPointKind])

Encapsulates parameter for one of entry points. It keeps entrypoint name and corresponding argument serialized.

In Haskell world, we keep an invariant of that contained value relates to one of entry points from entries list.

Constructors

UParamUnsafe (MText, ByteString)

type EntryPointKind = (Symbol, Type)

An entrypoint is described by two types: its name and type of argument.

type (?:) (n :: Symbol) (a :: k) = '(n, a)

A convenient alias for type-level name-something pair.

Construction

mkUParam :: (NicePackedValue a, LookupEntryPoint name entries ~ a, RequireUniqueEntryPoints entries) => Label name -> a -> UParam entries

Construct a UParam safely.

Deconstruction

data ConstrainedSome (c :: Type -> Constraint) where

This type can store any value that satisfies a certain constraint.

Constructors

ConstrainedSome :: c a => a -> ConstrainedSome c

class UnpackUParam (c :: Type -> Constraint) entries

This class is needed to implement unpackUParam.

Minimal complete definition

unpackUParam

unpackUParam :: UnpackUParam c entries => UParam entries -> Either EntryPointLookupError (MText, ConstrainedSome c)

Turn UParam into a Haskell value. Since we don't know its type in compile time, we have to erase it using ConstrainedSome. The user of this function can require arbitrary constraint to hold (depending on how they want to use the result).

Casting to homomorphic value

type SomeInterface = '['("SomeEntrypoints", Void)]

Pseudo value for UParam type variable.

type UParam_ = UParam SomeInterface

Homomorphic version of UParam, forgets the exact interface.

Pattern-matching

type EntryPointsImpl inp out entries = Rec (CaseClauseU inp out) entries

Implementations of some entry points.

Note that this thing inherits properties of Rec, e.g. you can Data.Vinyl.Core.rappend implementations for two entrypoint sets when assembling scattered parts of a contract.

type UParamFallback inp out = ((MText, ByteString) ': inp) :-> out

An action invoked when user-provided entrypoint is not found.

data EntryPointLookupError

Constructors

NoSuchEntryPoint MText

ArgumentUnpackFailed

class CaseUParam (entries :: [EntryPointKind])

Make up a "case" over entry points.

Minimal complete definition

caseUParamUnsafe

caseUParam :: (CaseUParam entries, RequireUniqueEntryPoints entries) => Rec (CaseClauseU inp out) entries -> UParamFallback inp out -> (UParam entries ': inp) :-> out

Pattern-match on given UParam entries.

You have to provide all case branches and a fallback action on case when entrypoint is not found.

caseUParamT :: forall entries inp out clauses. (clauses ~ Rec (CaseClauseU inp out) entries, RecFromTuple clauses, CaseUParam entries) => IsoRecTuple clauses -> UParamFallback inp out -> (UParam entries ': inp) :-> out

Like caseUParam, but accepts a tuple of clauses, not a Rec.

uparamFallbackFail :: UParamFallback inp out

Default implementation for UParamFallback, simply reports an error.

Constraints

type family LookupEntryPoint (name :: Symbol) (entries :: [EntryPointKind]) :: Type where ...

Get type of entrypoint argument by its name.

Equations

LookupEntryPoint name ('(name, a) ': _) = a

LookupEntryPoint name (_ ': entries) = LookupEntryPoint name entries

LookupEntryPoint name '[] = TypeError ((Text "Entry point " :<>: ShowType name) :<>: Text " in not in the entry points list")

type family RequireUniqueEntryPoints (entries :: [EntryPointKind]) :: Constraint where ...

Ensure that given entry points do no contain duplicated names.

Equations

RequireUniqueEntryPoints entries = RequireAllUnique "entrypoint" (Eval (Map Fst entries))

Conversion from ADT

uparamFromAdt :: UParamLinearize up => up -> UParam (UParamLinearized up)

Make up UParam from ADT sum.

Entry points template will consist of (constructorName, constructorFieldType) pairs. Each constructor is expected to have exactly one field.

type UParamLinearize p = (Generic p, GUParamLinearize (Rep p))

Constraint required by uparamFromAdt.

type UParamLinearized p = GUParamLinearized (Rep p)

Entry points template derived from given ADT sum.

Documentation

pbsUParam :: forall ctorName. KnownSymbol ctorName => ParamBuildingStep

Note that calling given entrypoints involves constructing UParam.

Internals used for entrypoint-wise migrations

unwrapUParam :: (UParam entries ': s) :-> ((MText, ByteString) ': s)

Helper instruction which extracts content of UParam.

Orphan instances

CustomErrorHasDoc "uparamArgumentUnpackFailed"

Instance details

Methods

customErrDocMdCause :: Markdown

customErrDocMdCauseInEntrypoint :: Markdown

customErrClass :: ErrorClass

customErrArgumentSemantics :: Maybe Markdown

CustomErrorHasDoc "uparamNoSuchEntryPoint"

Instance details

Methods

customErrDocMdCause :: Markdown

customErrDocMdCauseInEntrypoint :: Markdown

customErrClass :: ErrorClass

customErrArgumentSemantics :: Maybe Markdown

Buildable (CustomError "uparamArgumentUnpackFailed")

Instance details

Methods

build :: CustomError "uparamArgumentUnpackFailed" -> Builder

Buildable (CustomError "uparamNoSuchEntryPoint")

Instance details

Methods

build :: CustomError "uparamNoSuchEntryPoint" -> Builder