Lorentz.UStore.Types

Description

UStore definition and common type-level stuff.

UStore and related type definitions

newtype UStore (a :: Type)

Gathers multple fields and BigMaps under one object.

Type argument of this datatype stands for a "store template" - a datatype with one constructor and multiple fields, each containing an object of type UStoreFieldExt or |~> and corresponding to single virtual field or BigMap respectively. It's also possible to parameterize it with a larger type which is a product of types satisfying the above property.

Constructors

UStore

Fields

unUStore :: BigMap ByteString ByteString

newtype k |~> v

Describes one virtual big map in the storage.

Constructors

UStoreSubMap

Fields

unUStoreSubMap :: Map k v

newtype UStoreFieldExt (m :: UStoreMarkerType) (v :: Type)

Describes plain field in the storage.

Constructors

UStoreField

Fields

unUStoreField :: v

type UStoreField = UStoreFieldExt UMarkerPlainField

Just a plain field used as data.

type UStoreMarkerType = UStoreMarker -> Type

Specific kind used to designate markers for UStoreFieldExt.

We suggest that fields may serve different purposes and so annotated with special markers accordingly. See example below.

This kind is implemented like that because we want markers to differ from all other types in kind; herewith UStoreMarkerType is still an open kind (has potentially infinite number of inhabitants).

data UMarkerPlainField :: UStoreMarkerType

Extras

class KnownUStoreMarker (marker :: UStoreMarkerType) where

Allows to specify format of key under which fields of this type are stored. Useful to avoid collisions.

Minimal complete definition

Nothing

Associated Types

type ShowUStoreField marker v :: ErrorMessage

Display type-level information about UStore field with given marker and field value type. Used for error messages.

Methods

mkFieldMarkerUKey :: MText -> ByteString

By field name derive key under which field should be stored.

mkFieldMarkerUKey :: MText -> ByteString

By field name derive key under which field should be stored.

mkFieldMarkerUKeyL :: forall marker field. KnownUStoreMarker marker => Label field -> ByteString

Version of mkFieldMarkerUKey which accepts label.

mkFieldUKey :: forall (store :: Type) field. KnownUStoreMarker (GetUStoreFieldMarker store field) => Label field -> ByteString

Shortcut for mkFieldMarkerUKey which accepts not marker but store template and name of entry.

type UStoreSubmapKey k = (MText, k)

What do we serialize when constructing big_map key for accessing an UStore submap.

type UStoreSubmapKeyT k = TPair (ToT MText) k

Type-lookup-by-name

type GetUStoreKey store name = MSKey (GetUStore name store)

Get type of submap key.

type GetUStoreValue store name = MSValue (GetUStore name store)

Get type of submap value.

type GetUStoreField store name = FSValue (GetUStore name store)

Get type of plain field. This ignores marker with field type.

type GetUStoreFieldMarker store name = FSMarker (GetUStore name store)

Get kind of field.

Marked fields

type PickMarkedFields marker template = GPickMarkedFields marker (Rep template)

Collect all fields with the given marker.

Internals

data ElemSignature

What was found on lookup by constructor name.

This keeps either type arguments of |~> or UStoreFieldExt.

Constructors

MapSignature Type Type

FieldSignature UStoreMarkerType Type

type GetUStore name a = MERequireFound name a (GLookupStore name (Rep a))

Get map signature from the constructor with a given name.

type family MSKey (ms :: ElemSignature) :: Type where ...

Equations

MSKey (MapSignature k _) = k

MSKey (FieldSignature _ _) = TypeError (Text "Expected UStore submap, but field was referred")

type family MSValue (ms :: ElemSignature) :: Type where ...

Equations

MSValue (MapSignature _ v) = v

MSValue (FieldSignature _ _) = TypeError (Text "Expected UStore submap, but field was referred")

type family FSValue (ms :: ElemSignature) :: Type where ...

Equations

FSValue (FieldSignature _ v) = v

FSValue (MapSignature _ _) = TypeError (Text "Expected UStore field, but submap was referred")

type family FSMarker (ms :: ElemSignature) :: UStoreMarkerType where ...

Equations

FSValue (FieldSignature m _) = m

FSValue (MapSignature _ _) = TypeError (Text "Expected UStore field, but submap was referred")