Lorentz.Zip

Description

Stack zipping.

This module provides functions for flattening stacks into tuples.

Also here we define an instance which turns any instruction, not only lambdas, into a valid value.

Documentation

class ZipInstr (s :: [Type]) where

Zipping stack into tuple and back.

Associated Types

type ZippedStack s :: Type

A type which contains the whole stack zipped.

Methods

zipInstr :: s :-> '[ZippedStack s]

Fold given stack into single value.

unzipInstr :: '[ZippedStack s] :-> s

Unfold given stack from a single value.

type ZipInstrs ss = Each '[ZipInstr] ss

Require several stacks to comply ZipInstr constraint.

zippingStack :: ZipInstrs [inp, out] => (inp :-> out) -> Lambda (ZippedStack inp) (ZippedStack out)

Flatten both ends of instruction stack.

unzippingStack :: ZipInstrs [inp, out] => Lambda (ZippedStack inp) (ZippedStack out) -> inp :-> out

Unflatten both ends of instruction stack.

Orphan instances

(ZipInstr inp, ZipInstr out) => IsoValue (inp :-> out)

Instance details

Associated Types

type ToT (inp :-> out) :: T

Methods

toVal :: (inp :-> out) -> Value (ToT (inp :-> out))

fromVal :: Value (ToT (inp :-> out)) -> inp :-> out

(HasTypeAnn (ZippedStack i), HasTypeAnn (ZippedStack o)) => HasTypeAnn (i :-> o)

Instance details

Methods

getTypeAnn :: Notes (ToT (i :-> o))